+++ guildSettings.py
@@ -7,8 +7,12 @@ _SETTINGS_DEFAULTS = {
     "restrictToChannel" : None,
     "restrictToRoles" : [],
     "restrictToRolesInverted" : False,
-    "usageCooldown" : 0
+    "usageCooldown" : 0,
+    "antispamEnabled" : True,
+    "antispamAlertChannel" : None
 }
+propertyTypes = list|bool|None|int
+_guildSettings:dict[int,dict[str,propertyTypes]]
 
 def _saveSettings() -> None:
 
@@ -18,7 +22,19 @@ def _saveSettings() -> None:
 def _createGuildDefaults(guildId:int) -> None:
 
     if _guildSettings.get(guildId) is None:
-        _guildSettings[guildId] = {k : (v.copy() if type(v) in (list,dict) else v) for k,v in _SETTINGS_DEFAULTS.items()}
+        _guildSettings[guildId] = {}
+
+    curGuildSettings = _guildSettings[guildId]
+
+    defaultObj = object()
+    for key,value in _SETTINGS_DEFAULTS.items():
+        if curGuildSettings.get(key,defaultObj) is defaultObj:
+            if type(value) in (list,dict):
+                toSetValue = value.copy()
+            else:
+                toSetValue = value
+            curGuildSettings[key] = toSetValue
+
     _saveSettings()
 
 def _load() -> None:
@@ -40,13 +56,13 @@ def _load() -> None:
 _load()
 
 # doesn't need to be async but kept just in case
-async def setGuildSetting(guildId:int,property:str,value:list|bool|None|int) -> None:
+async def setGuildSetting(guildId:int,property:str,value:propertyTypes) -> None:
 
     _createGuildDefaults(guildId)
     _guildSettings[guildId][property] = value
     _saveSettings()
 
-async def getGuildSettings(guildId:int) -> dict[str,list|bool|None|int]:
+async def getGuildSettings(guildId:int) -> dict[str,propertyTypes]:
 
     _createGuildDefaults(guildId)
     return _guildSettings[guildId]